+++ amazon/cognito/fake/src/main/kotlin/org/http4k/connect/amazon/cognito/oauth/CognitoPoolClientValidator.kt
@@ -1,6 +1,7 @@
 package org.http4k.connect.amazon.cognito.oauth
 
 import org.http4k.connect.amazon.cognito.CognitoPool
+import org.http4k.connect.amazon.cognito.model.ClientSecret
 import org.http4k.connect.storage.Storage
 import org.http4k.core.Request
 import org.http4k.core.Uri
@@ -12,12 +13,16 @@ class CognitoPoolClientValidator(private val storage: Storage<CognitoPool>) : Cl
         storage.hasAppClient(clientId)
 
     override fun validateCredentials(request: Request, clientId: ClientId, clientSecret: String) =
-        storage.hasAppClient(clientId)
+        storage.hasAppClient(clientId, ClientSecret.of(clientSecret))
 
     override fun validateRedirection(request: Request, clientId: ClientId, redirectionUri: Uri) = true
 
     override fun validateScopes(request: Request, clientId: ClientId, scopes: List<String>) = true
 }
 
-private fun Storage<CognitoPool>.hasAppClient(clientId: ClientId) =
-    keySet().any { this[it]!!.clients.any { it.ClientId.value == clientId.value } }
+private fun Storage<CognitoPool>.hasAppClient(clientId: ClientId, secret: ClientSecret? = null) =
+    keySet().any {
+        this[it]!!.clients.any {
+            it.ClientId.value == clientId.value && (secret == null || it.ClientSecret == secret)
+        }
+    }