+++ amazon/cognito/fake/src/main/kotlin/org/http4k/connect/amazon/cognito/registerOAuthClient.kt
@@ -19,7 +19,7 @@ fun FakeCognito.registerOAuthClient(id: ClientId): Credentials =
         createUserPoolDomain(poolId, CloudFrontDomain.of(id.value)).valueOrNull()
         val userPoolClient = createUserPoolClient(
             UserPoolId = poolId,
-            ClientName = ClientName.of(id.value),
+            ClientName = ClientName.of(id.value)
         ).valueOrNull()!!.UserPoolClient
 
         Credentials(userPoolClient.ClientId.value, userPoolClient.ClientSecret!!.value)