+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This list is not intended to be all-encompassing - it will document major and br
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
 ### v3.26.1.0 (uncut)
-- **http4k-connect-amazon-dynamodb** - Add tableMapper batchGet and batchDelete operations
+- **http4k-connect-amazon-dynamodb** - Add tableMapper batchGet and batchDelete operations. H/T @oharaandrew314
+- **http4k-connect-cognito-fake** - Enforce matching of Client Secret as well as ClientId 
 
 ### v3.26.0.0
 - **http4k-connect-*** - Upgrade dependencies.